Don't relink if the symlink is already correct.

+++ docker-link-exports.sh
@@ -38,10 +38,15 @@ if [ -d ${DATA_EXPORT_DIR} ]; then
 
         echo -n "Symlinking ${target} into place... "
         if [ -e ${source} ]; then
-            # Unlink the source and symlink the target in.
-            source_tmp=$(mktemp -u $(dirname ${source})/$(basename ${source}).XXXX)
-            mv -f ${source} ${source_tmp}
-            ln -s $(relpath ${target} ${source_dir}) ${source} && rm -rf ${source_tmp}
+            # Don't do anything if the symlink is already in place.
+            if [ -L ${source} -a $(readlink -e ${source}) = ${target} ]; then
+                echo -n "already "
+            else
+                # Unlink the source and symlink the target in.
+                source_tmp=$(mktemp -u $(dirname ${source})/$(basename ${source}).XXXX)
+                mv -f ${source} ${source_tmp}
+                ln -s $(relpath ${target} ${source_dir}) ${source} && rm -rf ${source_tmp}
+            fi
         else
             ln -s $(relpath ${target} ${source_dir}) ${source}
         fi
